feat(users): implementa gestión de usuarios con CRUD completo
- Agrega UserRepository con métodos de acceso a datos (get_all, get_by_id, create, update, soft_delete)
- Implementa UserService con validaciones de email único y privilegios admin
- Crea UserReadSerializer (excluye password) y UserWriteSerializer (write_only password)
- Agrega UserViewSet con endpoints REST, soft-delete y filtros de búsqueda
- Configura permisos IsAdminUser para acceso solo admin/staff
- Registra rutas en apps.users.urls y config.urls
- Actualiza AUTH_USER_MODEL en settings.py

Refer #18
