+++ frontend/src/App.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
-import Login from './features/users/pages/Login';
-import DashboardLayout from './core/layout/DashboardLayout/DashboardLayout';
-import PosDashboard from './features/pos/pages/PosDashboard';
-import InventoryDashboard from './features/inventory/pages/InventoryDashboard';
-import HrDashboard from './features/hr/pages/HrDashboard';
-import UsersDashboard from './features/users/pages/UsersDashboard';
+import Login from '~users/pages/Login';
+import DashboardLayout from '~core/layout/DashboardLayout/DashboardLayout';
+import PosDashboard from '~pos/pages/PosDashboard';
+import InventoryDashboard from '~inventory/pages/InventoryDashboard';
+import HrDashboard from '~hr/pages/HrDashboard';
+import UsersDashboard from '~users/pages/UsersDashboard';
 
 function App() {
   return (
