+++ frontend/src/features/users/components/UserFormModal/index.jsx
@@ -1,7 +1,4 @@
-import React, { useEffect } from 'react';
-import { useForm } from 'react-hook-form';
-import { zodResolver } from '@hookform/resolvers/zod';
-import * as z from 'zod';
+import React from 'react';
 import {
     Dialog,
     DialogContent,
@@ -13,121 +10,19 @@ import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Switch } from '@/components/ui/switch';
-import { userFrontendService } from '../services/userFrontendService';
-
-const baseSchema = z.object({
-    email: z.string().email('Debe ser un correo válido'),
-    first_name: z.string().min(1, 'El nombre es requerido'),
-    last_name: z.string().min(1, 'El apellido es requerido'),
-    phone: z.string().optional(),
-    is_staff: z.boolean().default(false),
-});
-
-const createUserSchema = baseSchema.extend({
-    password: z.string().min(6, 'La contraseña debe tener al menos 6 caracteres'),
-    confirm_password: z.string()
-}).refine((data) => data.password === data.confirm_password, {
-    message: "Las contraseñas no coinciden",
-    path: ["confirm_password"],
-});
-
-const editUserSchema = baseSchema.extend({
-    password: z.string().optional(),
-    confirm_password: z.string().optional()
-}).refine((data) => {
-    if (data.password && data.password !== data.confirm_password) {
-        return false;
-    }
-    return true;
-}, {
-    message: "Las contraseñas no coinciden",
-    path: ["confirm_password"],
-});
+import { useUserFormModal } from './useUserFormModal';
 
 const UserFormModal = ({ isOpen, onClose, user }) => {
-    const isEditing = !!user;
-    const schema = isEditing ? editUserSchema : createUserSchema;
-
     const {
         register,
         handleSubmit,
         setValue,
-        watch,
-        reset,
-        setError,
-        formState: { errors, isSubmitting }
-    } = useForm({
-        resolver: zodResolver(schema),
-        defaultValues: {
-            email: '',
-            first_name: '',
-            last_name: '',
-            phone: '',
-            is_staff: false,
-            password: '',
-            confirm_password: '',
-        }
-    });
-
-    const isStaffValue = watch('is_staff');
-
-    useEffect(() => {
-        if (isOpen) {
-            if (isEditing) {
-                reset({
-                    email: user.email || '',
-                    first_name: user.first_name || '',
-                    last_name: user.last_name || '',
-                    phone: user.phone || '',
-                    is_staff: user.is_staff || false,
-                    password: '',
-                    confirm_password: '',
-                });
-            } else {
-                reset({
-                    email: '',
-                    first_name: '',
-                    last_name: '',
-                    phone: '',
-                    is_staff: false,
-                    password: '',
-                    confirm_password: '',
-                });
-            }
-        }
-    }, [isOpen, isEditing, user, reset]);
-
-    const onSubmit = async (data) => {
-        try {
-            const payload = { ...data };
-            // Remove confirm_password as backend doesn't need it
-            delete payload.confirm_password;
-
-            // If editing and password is empty, don't send it
-            if (isEditing && !payload.password) {
-                delete payload.password;
-            }
-
-            if (isEditing) {
-                // Remove email when editing, as it shouldn't be updated
-                delete payload.email;
-                await userFrontendService.updateUser(user.id, payload);
-            } else {
-                await userFrontendService.createUser(payload);
-            }
-            onClose(true); // true indicates successful save
-        } catch (error) {
-            // Handle 400 error from backend (like email already exists)
-            console.error('Save failed', error);
-            if (error.response?.data?.email) {
-                setError('email', { type: 'server', message: 'Este correo ya está registrado' });
-            } else if (error.response?.data?.detail) {
-                setError('root', { type: 'server', message: error.response.data.detail });
-            } else {
-                setError('root', { type: 'server', message: 'Ocurrió un error inesperado al guardar.' });
-            }
-        }
-    };
+        isStaffValue,
+        errors,
+        isSubmitting,
+        isEditing,
+        onSubmit
+    } = useUserFormModal(isOpen, onClose, user);
 
     return (
         <Dialog open={isOpen} onOpenChange={(open) => !open && onClose(false)}>
