+++ frontend/vite.config.js
@@ -9,6 +9,11 @@ export default defineConfig({
   resolve: {
     alias: {
       '@': path.resolve(__dirname, './src'),
+      '~core': path.resolve(__dirname, './src/core'),
+      '~users': path.resolve(__dirname, './src/features/users'),
+      '~pos': path.resolve(__dirname, './src/features/pos'),
+      '~inventory': path.resolve(__dirname, './src/features/inventory'),
+      '~hr': path.resolve(__dirname, './src/features/hr'),
     },
   },
 })
